requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
Flask-CORS==4.0.0
Flask-Mail==0.9.1
markdown2==2.4.12
Flask-Login==0.6.3
python-slugify==8.0.4 